+++ lib/core/configuration.ts
@@ -60,7 +60,7 @@ interface GlobalWatcherConfiguration {
 interface ProcessConfiguration {
   id: string
   cmd?: string
-  worker?: string
+  worker?: string[]
   env?: Record<string, string>
   cwd?: string
   path?: string
@@ -111,7 +111,7 @@ const ConfigurationSchema = z.object({
     z.object({
       id: z.string().min(1).max(64).regex(/^[a-z0-9@._\-]+$/i, "Process ID can only contain characters a-Z 0-9 . _ - or @"),
       cmd: z.optional(z.string()),
-      worker: z.optional(z.string()),
+      worker: z.optional(z.array(z.string())),
       cwd: z.optional(z.string()),
       env: z.optional(z.object({})),
       cluster: z.optional(z.object({